+++ meta_package_manager/managers/cargo.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 import re
 from typing import Iterator
 
-from click_extra.platforms import ALL_PLATFORMS
+from extra_platforms import ALL_PLATFORMS
 
 from meta_package_manager.base import Package, PackageManager
 from meta_package_manager.capabilities import (