+++ meta_package_manager/managers/pip.py
@@ -21,7 +21,7 @@ import re
 from functools import cached_property
 from typing import TYPE_CHECKING, Iterator
 
-from click_extra.platforms import ALL_PLATFORMS
+from extra_platforms import ALL_PLATFORMS
 
 from meta_package_manager.base import Package, PackageManager
 from meta_package_manager.capabilities import (