+++ meta_package_manager/managers/scoop.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 import re
 from typing import Iterator
 
-from click_extra.platforms import WINDOWS
+from extra_platforms import WINDOWS
 
 from meta_package_manager.base import Package, PackageManager
 from meta_package_manager.capabilities import (