+++ meta_package_manager/managers/steamcmd.py
@@ -16,7 +16,7 @@
 
 from __future__ import annotations
 
-from click_extra.platforms import ALL_PLATFORMS
+from extra_platforms import ALL_PLATFORMS
 
 from meta_package_manager.base import PackageManager
 from meta_package_manager.capabilities import version_not_implemented