+++ meta_package_manager/managers/zypper.py
@@ -21,7 +21,7 @@ from operator import itemgetter
 from typing import Iterator, TypedDict
 
 import xmltodict
-from click_extra.platforms import UNIX_WITHOUT_MACOS
+from extra_platforms import UNIX_WITHOUT_MACOS
 
 from meta_package_manager.base import Package, PackageManager
 from meta_package_manager.capabilities import version_not_implemented