+++ tests/test_cli.py
@@ -24,8 +24,8 @@ from typing import Collection, Iterator
 
 import pytest
 from boltons.strutils import strip_ansi
-from click_extra.platforms import is_windows
 from click_extra.tabulate import output_formats
+from extra_platforms import is_windows
 
 from meta_package_manager import __version__
 from meta_package_manager.pool import pool