+++ tests/test_cli_backup.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 
 import pytest
 from boltons.strutils import strip_ansi
-from click_extra.platforms import is_windows
+from extra_platforms import is_windows
 
 from meta_package_manager.base import Operations