+++ tests/test_manager_apt.py
@@ -20,7 +20,7 @@ from __future__ import annotations
 import subprocess
 
 import pytest
-from click_extra.pytest import unless_linux
+from extra_platforms.pytest import unless_linux
 
 
 @pytest.fixture